AirIndia-RAG-IntelligenceBOT/requirements.txt

boto3==1.34.62
langchain-community==0.0.31
langchain-text-splitters==0.0.1
langchain-chroma
tiktoken
streamlit==1.32.2
python-dotenv
argparse
langchain-huggingface
sentence-transformers
ollama
pytest
pytest-sugar
pytest-html
pytest-csv
ragas
datasets
pandas
